requirements.txt
fabric
jinja2 
requests
bitbucket-api